Fix aur_search
the current implementation of aur_search states:

Searches for packages by a query, optionally filtering by maintainer
name.

This is untrue as &maintainer is not a valid argument. Instead
&by=maintainer should be used to search for a maintainer with the &arg=
becoming the maintainer to search for.

Additionally an empty query will return orphans. As appending "&arg="
still counts as an empty query, drop the Option wrapper and make query
a &str letting the user specify "" for orphans.

This patch implements &by= for all supported types. And renames the
method to aur_search_by.