Fix search (#3)
* Fix aur_search

the current implementation of aur_search states:

Searches for packages by a query, optionally filtering by maintainer
name.

This is untrue as &maintainer is not a valid argument. Instead
&by=maintainer should be used to search for a maintainer with the &arg=
becoming the maintainer to search for.

Additionally an empty query will return orphans. As appending "&arg="
still counts as an empty query, drop the Option wrapper and make query
a &str letting the user specify "" for orphans.

This patch implements &by= for all supported types. And renames the
method to aur_search_by.

* Add aur_orphans and aur_search

The aur_orphans method returns all orphans in the AUR. This is a convenience
for aur_search("", SearchBy::Maintainer).

The aur_search method always searches by name-desc. This is what the RPC
searches by if &by= is not specified.